+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Correctly resolve paths relative to redirected modules
+
 ## [0.1.1] - 2025-11-25
 
 ### Fixed